+++ test/TBD/Inputs/struct.log
@@ -1,0 +1,42 @@
+<unknown>:0: error: symbol '_T04test10PropertiesV9publicLetSivfi' (test.Properties.(publicLet : Swift.Int).(variable initialization expression)) is in generated IR file, but not in TBD file
+<unknown>:0: error: symbol '_T04test10PropertiesV9publicVarSivfi' (test.Properties.(publicVar : Swift.Int).(variable initialization expression)) is in generated IR file, but not in TBD file
+<unknown>:0: error: symbol '_T04test11PrivateInitVMa' (type metadata accessor for test.PrivateInit) is in generated IR file, but not in TBD file
+<unknown>:0: error: symbol '_T04test11PrivateInitVMn' (nominal type descriptor for test.PrivateInit) is in generated IR file, but not in TBD file
+<unknown>:0: error: symbol '_T04test11PrivateInitVN' (type metadata for test.PrivateInit) is in generated IR file, but not in TBD file
+<unknown>:0: error: symbol '_T04test11PrivateInitVWV' (value witness table for test.PrivateInit) is in generated IR file, but not in TBD file
+<unknown>:0: error: symbol '_T04test14PrivateMethodsVMa' (type metadata accessor for test.PrivateMethods) is in generated IR file, but not in TBD file
+<unknown>:0: error: symbol '_T04test14PrivateMethodsVMn' (nominal type descriptor for test.PrivateMethods) is in generated IR file, but not in TBD file
+<unknown>:0: error: symbol '_T04test14PrivateMethodsVN' (type metadata for test.PrivateMethods) is in generated IR file, but not in TBD file
+<unknown>:0: error: symbol '_T04test14PrivateMethodsVWV' (value witness table for test.PrivateMethods) is in generated IR file, but not in TBD file
+<unknown>:0: error: symbol '_T04test14PrivateNothingVMa' (type metadata accessor for test.PrivateNothing) is in generated IR file, but not in TBD file
+<unknown>:0: error: symbol '_T04test14PrivateNothingVMn' (nominal type descriptor for test.PrivateNothing) is in generated IR file, but not in TBD file
+<unknown>:0: error: symbol '_T04test14PrivateNothingVN' (type metadata for test.PrivateNothing) is in generated IR file, but not in TBD file
+<unknown>:0: error: symbol '_T04test14PrivateNothingVWV' (value witness table for test.PrivateNothing) is in generated IR file, but not in TBD file
+<unknown>:0: error: symbol '_T04test14PrivateStaticsV10privateLetSivZ' (static test.PrivateStatics.privateLet : Swift.Int) is in generated IR file, but not in TBD file
+<unknown>:0: error: symbol '_T04test14PrivateStaticsV10privateVarSivZ' (static test.PrivateStatics.privateVar : Swift.Int) is in generated IR file, but not in TBD file
+<unknown>:0: error: symbol '_T04test14PrivateStaticsVMa' (type metadata accessor for test.PrivateStatics) is in generated IR file, but not in TBD file
+<unknown>:0: error: symbol '_T04test14PrivateStaticsVMn' (nominal type descriptor for test.PrivateStatics) is in generated IR file, but not in TBD file
+<unknown>:0: error: symbol '_T04test14PrivateStaticsVN' (type metadata for test.PrivateStatics) is in generated IR file, but not in TBD file
+<unknown>:0: error: symbol '_T04test14PrivateStaticsVWV' (value witness table for test.PrivateStatics) is in generated IR file, but not in TBD file
+<unknown>:0: error: symbol '_T04test17PrivatePropertiesVMa' (type metadata accessor for test.PrivateProperties) is in generated IR file, but not in TBD file
+<unknown>:0: error: symbol '_T04test17PrivatePropertiesVMn' (nominal type descriptor for test.PrivateProperties) is in generated IR file, but not in TBD file
+<unknown>:0: error: symbol '_T04test17PrivatePropertiesVN' (type metadata for test.PrivateProperties) is in generated IR file, but not in TBD file
+<unknown>:0: error: symbol '_T04test17PrivatePropertiesVWV' (value witness table for test.PrivateProperties) is in generated IR file, but not in TBD file
+<unknown>:0: error: symbol '_T04test7StaticsV10privateVarSivZ' (static test.Statics.privateVar : Swift.Int) is in generated IR file, but not in TBD file
+<unknown>:0: error: symbol '_T04test7StaticsV9publicVarSifau' (test.Statics.publicVar.unsafeMutableAddressor : Swift.Int) is in generated IR file, but not in TBD file
+<unknown>:0: error: symbol '_T04test7StaticsV9publicVarSivZ' (static test.Statics.publicVar : Swift.Int) is in generated IR file, but not in TBD file
+<unknown>:0: error: symbol '_T04test10PropertiesV10privateLetSifau' (test.Properties.privateLet.unsafeMutableAddressor : Swift.Int) is in TBD file, but not in generated IR
+<unknown>:0: error: symbol '_T04test10PropertiesV10privateLetSiv' (test.Properties.privateLet : Swift.Int) is in TBD file, but not in generated IR
+<unknown>:0: error: symbol '_T04test10PropertiesV15publicVarGetSetSifm' (test.Properties.publicVarGetSet.materializeForSet : Swift.Int) is in TBD file, but not in generated IR
+<unknown>:0: error: symbol '_T04test10PropertiesV9publicLetSifau' (test.Properties.publicLet.unsafeMutableAddressor : Swift.Int) is in TBD file, but not in generated IR
+<unknown>:0: error: symbol '_T04test10PropertiesV9publicLetSifg' (test.Properties.publicLet.getter : Swift.Int) is in TBD file, but not in generated IR
+<unknown>:0: error: symbol '_T04test10PropertiesV9publicLetSiv' (test.Properties.publicLet : Swift.Int) is in TBD file, but not in generated IR
+<unknown>:0: error: symbol '_T04test10PropertiesV9publicVarSifg' (test.Properties.publicVar.getter : Swift.Int) is in TBD file, but not in generated IR
+<unknown>:0: error: symbol '_T04test10PropertiesV9publicVarSifm' (test.Properties.publicVar.materializeForSet : Swift.Int) is in TBD file, but not in generated IR
+<unknown>:0: error: symbol '_T04test10PropertiesV9publicVarSifs' (test.Properties.publicVar.setter : Swift.Int) is in TBD file, but not in generated IR
+<unknown>:0: error: symbol '_T04test7StaticsV10privateLetSifau' (test.Statics.privateLet.unsafeMutableAddressor : Swift.Int) is in TBD file, but not in generated IR
+<unknown>:0: error: symbol '_T04test7StaticsV15publicVarGetSetSifmZ' (static test.Statics.publicVarGetSet.materializeForSet : Swift.Int) is in TBD file, but not in generated IR
+<unknown>:0: error: symbol '_T04test7StaticsV9publicLetSifgZ' (static test.Statics.publicLet.getter : Swift.Int) is in TBD file, but not in generated IR
+<unknown>:0: error: symbol '_T04test7StaticsV9publicVarSifgZ' (static test.Statics.publicVar.getter : Swift.Int) is in TBD file, but not in generated IR
+<unknown>:0: error: symbol '_T04test7StaticsV9publicVarSifmZ' (static test.Statics.publicVar.materializeForSet : Swift.Int) is in TBD file, but not in generated IR
+<unknown>:0: error: symbol '_T04test7StaticsV9publicVarSifsZ' (static test.Statics.publicVar.setter : Swift.Int) is in TBD file, but not in generated IR
